+++ src/aks-preview/azext_aks_preview/_params.py
@@ -25,7 +25,8 @@ from ._validators import (
     validate_nodepool_labels, validate_vnet_subnet_id)
 from ._consts import CONST_OUTBOUND_TYPE_LOAD_BALANCER, \
     CONST_OUTBOUND_TYPE_USER_DEFINED_ROUTING, CONST_SCALE_SET_PRIORITY_REGULAR, CONST_SCALE_SET_PRIORITY_SPOT, \
-    CONST_SPOT_EVICTION_POLICY_DELETE, CONST_SPOT_EVICTION_POLICY_DEALLOCATE
+    CONST_SPOT_EVICTION_POLICY_DELETE, CONST_SPOT_EVICTION_POLICY_DEALLOCATE, \
+    CONST_NODEPOOL_MODE_SYSTEM, CONST_NODEPOOL_MODE_USER
 
 
 def load_arguments(self, _):
@@ -92,6 +93,7 @@ def load_arguments(self, _):
         c.argument('enable_vmss', action='store_true', help='To be deprecated. Use vm_set_type instead.')
         c.argument('vm_set_type', type=str, validator=validate_vm_set_type)
         c.argument('node_zones', zones_type, options_list=['--node-zones', '--zones', '-z'], help='(--node-zones will be deprecated, use --zones) Space-separated list of availability zones where agent nodes will be placed.')
+        c.argument('enable_node_public_ip', action='store_true', is_preview=True)
         c.argument('enable_pod_security_policy', action='store_true')
         c.argument('node_resource_group')
         c.argument('attach_acr', acr_arg_type)
@@ -133,6 +135,7 @@ def load_arguments(self, _):
             c.argument('nodepool_name', type=str, options_list=['--name', '-n'], validator=validate_nodepool_name, help='The node pool name.')
             c.argument('tags', tags_type)
             c.argument('node_zones', zones_type, options_list=['--node-zones', '--zones', '-z'], help='(--node-zones will be deprecated) Space-separated list of availability zones where agent nodes will be placed.')
+            c.argument('enable_node_public_ip', action='store_true', is_preview=True)
             c.argument('node_vm_size', options_list=['--node-vm-size', '-s'], completer=get_vm_size_completion_list)
             c.argument('max_pods', type=int, options_list=['--max-pods', '-m'], validator=validate_max_pods)
             c.argument('os_type', type=str)
@@ -142,6 +145,7 @@ def load_arguments(self, _):
             c.argument('eviction_policy', arg_type=get_enum_type([CONST_SPOT_EVICTION_POLICY_DELETE, CONST_SPOT_EVICTION_POLICY_DEALLOCATE]), validator=validate_eviction_policy)
             c.argument('spot_max_price', type=float, validator=validate_spot_max_price)
             c.argument('labels', nargs='*', validator=validate_nodepool_labels)
+            c.argument('mode', arg_type=get_enum_type([CONST_NODEPOOL_MODE_SYSTEM, CONST_NODEPOOL_MODE_USER]))
 
     for scope in ['aks nodepool show', 'aks nodepool delete', 'aks nodepool scale', 'aks nodepool upgrade', 'aks nodepool update']:
         with self.argument_context(scope) as c:
@@ -152,6 +156,7 @@ def load_arguments(self, _):
         c.argument('disable_cluster_autoscaler', options_list=["--disable-cluster-autoscaler", "-d"], action='store_true')
         c.argument('update_cluster_autoscaler', options_list=["--update-cluster-autoscaler", "-u"], action='store_true')
         c.argument('tags', tags_type)
+        c.argument('mode', arg_type=get_enum_type([CONST_NODEPOOL_MODE_SYSTEM, CONST_NODEPOOL_MODE_USER]))
 
     with self.argument_context('aks disable-addons') as c:
         c.argument('addons', options_list=['--addons', '-a'])