+++ src/spring-cloud/azext_spring_cloud/vendored_sdks/appplatform/operations/_apps_operations.py
@@ -113,9 +113,7 @@ class AppsOperations(object):
 
 
     def _create_or_update_initial(
-            self, resource_group_name, service_name, app_name, properties=None, custom_headers=None, raw=False, **operation_config):
-        app_resource = models.AppResource(properties=properties)
-
+            self, resource_group_name, service_name, app_name, app_resource, custom_headers=None, raw=False, **operation_config):
         # Construct URL
         url = self.create_or_update.metadata['url']
         path_format_arguments = {
@@ -167,7 +165,7 @@ class AppsOperations(object):
         return deserialized
 
     def create_or_update(
-            self, resource_group_name, service_name, app_name, properties=None, custom_headers=None, raw=False, polling=True, **operation_config):
+            self, resource_group_name, service_name, app_name, app_resource, custom_headers=None, raw=False, polling=True, **operation_config):
         """Create a new App or update an exiting App.
 
         :param resource_group_name: The name of the resource group that
@@ -178,8 +176,8 @@ class AppsOperations(object):
         :type service_name: str
         :param app_name: The name of the App resource.
         :type app_name: str
-        :param properties: Properties of the App resource
-        :type properties: ~azure.mgmt.appplatform.models.AppResourceProperties
+        :param app_resource: Parameters for the create or update operation
+        :type app_resource: ~azure.mgmt.appplatform.models.AppResource
         :param dict custom_headers: headers that will be added to the request
         :param bool raw: The poller return type is ClientRawResponse, the
          direct response alongside the deserialized response
@@ -197,7 +195,7 @@ class AppsOperations(object):
             resource_group_name=resource_group_name,
             service_name=service_name,
             app_name=app_name,
-            properties=properties,
+            app_resource=app_resource,
             custom_headers=custom_headers,
             raw=True,
             **operation_config
@@ -281,9 +279,7 @@ class AppsOperations(object):
 
 
     def _update_initial(
-            self, resource_group_name, service_name, app_name, properties=None, custom_headers=None, raw=False, **operation_config):
-        app_resource = models.AppResource(properties=properties)
-
+            self, resource_group_name, service_name, app_name, app_resource, custom_headers=None, raw=False, **operation_config):
         # Construct URL
         url = self.update.metadata['url']
         path_format_arguments = {
@@ -335,7 +331,7 @@ class AppsOperations(object):
         return deserialized
 
     def update(
-            self, resource_group_name, service_name, app_name, properties=None, custom_headers=None, raw=False, polling=True, **operation_config):
+            self, resource_group_name, service_name, app_name, app_resource, custom_headers=None, raw=False, polling=True, **operation_config):
         """Operation to update an exiting App.
 
         :param resource_group_name: The name of the resource group that
@@ -346,8 +342,8 @@ class AppsOperations(object):
         :type service_name: str
         :param app_name: The name of the App resource.
         :type app_name: str
-        :param properties: Properties of the App resource
-        :type properties: ~azure.mgmt.appplatform.models.AppResourceProperties
+        :param app_resource: Parameters for the update operation
+        :type app_resource: ~azure.mgmt.appplatform.models.AppResource
         :param dict custom_headers: headers that will be added to the request
         :param bool raw: The poller return type is ClientRawResponse, the
          direct response alongside the deserialized response
@@ -365,7 +361,7 @@ class AppsOperations(object):
             resource_group_name=resource_group_name,
             service_name=service_name,
             app_name=app_name,
-            properties=properties,
+            app_resource=app_resource,
             custom_headers=custom_headers,
             raw=True,
             **operation_config